+++ src/main.js
@@ -1,10 +1,12 @@
-import App from './App.vue'
+import App from '@/App.vue'
 import Vue from 'vue'
-import vuetify from './plugins/vuetify';
+import vuetify from '@/plugins/vuetify'
+import router from '@/router'
 
 Vue.config.productionTip = false
 
 new Vue({
+  router,
   vuetify,
   render: (h) => h(App)
 }).$mount('#app')
