+++ src/pages/Online.vue
@@ -1,0 +1,5 @@
+<template>
+    <v-app>
+    nnnnnnnnnnnnnnnnnnnn
+    </v-app>
+</template>
