+++ vue.config.js
@@ -1,7 +1,17 @@
+var path = require("path");
+
 module.exports = {
   "transpileDependencies": [
-    "vuetify"
+  "vuetify"
   ],
   publicPath: './',
-  assetsDir: ""
-}
+  assetsDir: "",
+  configureWebpack: {
+    resolve: {
+      alias: {
+        'vue$': 'vue/dist/vue.esm.js',
+        '@': path.resolve(__dirname, 'src/'), 
+      },
+    },
+  },
+};
